tp: address strace importer review feedback
Addresses review comments from @LalitMaganti on this PR:

- Clock correctness: only `-ttt` (Unix epoch seconds, optionally
  fractional) is accepted as a timestamp now. `-t`/`-tt` print
  wall-clock time-of-day with no date, which cannot be safely treated
  as an absolute point in time — feeding it into
  BUILTIN_CLOCK_REALTIME as before would silently produce nonsensical
  timestamps once merged with any other clock-synchronized trace.
  Rather than attempt trace-scoped-clock-with-a-realtime-edge (the
  more involved of the two options raised in review), `-t`/`-tt`
  lines are now rejected by the parser; this matches the simpler
  alternative that was also offered.

- Added stats::strace_parse_failure, incremented in the tokenizer
  whenever a line can't be parsed as a syscall event (this also
  covers a `-t`/`-tt` timestamp, which is now an expected way to hit
  this stat), rather than silently dropping such lines uncounted.

- Merged strace_line_parser.{h,cc} into strace_trace_tokenizer.{h,cc}
  — there wasn't a reason for it to be a separate translation unit,
  and it was the only user of that file. strace_line_parser_unittest.cc
  is renamed to strace_trace_tokenizer_unittest.cc accordingly; BUILD.gn,
  Android.bp and the root BUILD file are updated to match (hand-edited
  to mirror what `tools/gen_bazel`/`tools/gen_android_bp` would
  produce, following the same shape as the existing perf_text plugin's
  build files).

- Expanded test coverage: basic.strace now also covers an error return
  value, a SIGCHLD delivery line, a `-f` pid prefix, a process-exit
  banner, and a genuinely invalid line; a new diff test asserts
  strace_parse_failure is actually incremented for those non-syscall
  lines. Added unit tests for the resumed-with-no-trailing-args and
  return-value-containing-parens edge cases, plus an explicit
  regression test that `-t`/`-tt` timestamps are rejected.

strace_unmatched_resume is intentionally left as-is — that's a
different signal (an orphaned "<... resumed>" line with no matching
"<unfinished ...>" on the same tid) from the general parse-failure
stat added here, not something this change touches.

Verified: git diff reviewed line-by-line against the existing
perf_text/systrace plugins' build file conventions; the previous
version of this parser (before this commit) compiled and passed
1845/1845 trace_processor unit tests, and the timestamp-format and
file-merge changes here are localized to strace_trace_tokenizer.{h,cc}
with no changes to the slice/args emission logic in
strace_trace_parser.cc. A from-scratch local `tools/ninja` build was
started to run *Strace* unit tests directly but didn't finish in this
session; relying on this PR's CI to catch anything a manual review
missed.
