+++ Android.bp
@@ -3139,7 +3139,6 @@ cc_test {
         ":perfetto_src_trace_processor_plugins_stdlib_docs_stdlib_docs",
         ":perfetto_src_trace_processor_plugins_storage_tables_storage_tables",
         ":perfetto_src_trace_processor_plugins_strace_strace",
-        ":perfetto_src_trace_processor_plugins_strace_strace_line_parser",
         ":perfetto_src_trace_processor_plugins_string_functions_string_functions",
         ":perfetto_src_trace_processor_plugins_structural_tree_partition_structural_tree_partition",
         ":perfetto_src_trace_processor_plugins_symbolize_symbolize",
@@ -20087,19 +20086,11 @@ filegroup {
     ],
 }
 
-// GN: //src/trace_processor/plugins/strace:strace_line_parser
-filegroup {
-    name: "perfetto_src_trace_processor_plugins_strace_strace_line_parser",
-    srcs: [
-        "src/trace_processor/plugins/strace/strace_line_parser.cc",
-    ],
-}
-
 // GN: //src/trace_processor/plugins/strace:unittests
 filegroup {
     name: "perfetto_src_trace_processor_plugins_strace_unittests",
     srcs: [
-        "src/trace_processor/plugins/strace/strace_line_parser_unittest.cc",
+        "src/trace_processor/plugins/strace/strace_trace_tokenizer_unittest.cc",
     ],
 }
 
@@ -21038,7 +21029,6 @@ cc_library_static {
         ":perfetto_src_trace_processor_plugins_stdlib_docs_stdlib_docs",
         ":perfetto_src_trace_processor_plugins_storage_tables_storage_tables",
         ":perfetto_src_trace_processor_plugins_strace_strace",
-        ":perfetto_src_trace_processor_plugins_strace_strace_line_parser",
         ":perfetto_src_trace_processor_plugins_string_functions_string_functions",
         ":perfetto_src_trace_processor_plugins_structural_tree_partition_structural_tree_partition",
         ":perfetto_src_trace_processor_plugins_symbolize_symbolize",
@@ -23981,7 +23971,6 @@ cc_test {
         ":perfetto_src_trace_processor_plugins_stdlib_docs_stdlib_docs",
         ":perfetto_src_trace_processor_plugins_storage_tables_storage_tables",
         ":perfetto_src_trace_processor_plugins_strace_strace",
-        ":perfetto_src_trace_processor_plugins_strace_strace_line_parser",
         ":perfetto_src_trace_processor_plugins_strace_unittests",
         ":perfetto_src_trace_processor_plugins_string_functions_string_functions",
         ":perfetto_src_trace_processor_plugins_string_functions_unittests",
@@ -25102,7 +25091,6 @@ cc_library_static {
         ":perfetto_src_trace_processor_plugins_stdlib_docs_stdlib_docs",
         ":perfetto_src_trace_processor_plugins_storage_tables_storage_tables",
         ":perfetto_src_trace_processor_plugins_strace_strace",
-        ":perfetto_src_trace_processor_plugins_strace_strace_line_parser",
         ":perfetto_src_trace_processor_plugins_string_functions_string_functions",
         ":perfetto_src_trace_processor_plugins_structural_tree_partition_structural_tree_partition",
         ":perfetto_src_trace_processor_plugins_symbolize_symbolize",
@@ -25800,7 +25788,6 @@ cc_binary_host {
         ":perfetto_src_trace_processor_plugins_stdlib_docs_stdlib_docs",
         ":perfetto_src_trace_processor_plugins_storage_tables_storage_tables",
         ":perfetto_src_trace_processor_plugins_strace_strace",
-        ":perfetto_src_trace_processor_plugins_strace_strace_line_parser",
         ":perfetto_src_trace_processor_plugins_string_functions_string_functions",
         ":perfetto_src_trace_processor_plugins_structural_tree_partition_structural_tree_partition",
         ":perfetto_src_trace_processor_plugins_symbolize_symbolize",
