+++ BUILD
@@ -512,7 +512,6 @@ perfetto_cc_library(
         ":src_trace_processor_plugins_stdlib_docs_tables",
         ":src_trace_processor_plugins_storage_tables_storage_tables",
         ":src_trace_processor_plugins_strace_strace",
-        ":src_trace_processor_plugins_strace_strace_line_parser",
         ":src_trace_processor_plugins_string_functions_string_functions",
         ":src_trace_processor_plugins_structural_tree_partition_structural_tree_partition",
         ":src_trace_processor_plugins_structural_tree_partition_tables",
@@ -821,7 +820,6 @@ perfetto_cc_library(
         ":src_trace_processor_plugins_stdlib_docs_tables",
         ":src_trace_processor_plugins_storage_tables_storage_tables",
         ":src_trace_processor_plugins_strace_strace",
-        ":src_trace_processor_plugins_strace_strace_line_parser",
         ":src_trace_processor_plugins_string_functions_string_functions",
         ":src_trace_processor_plugins_structural_tree_partition_structural_tree_partition",
         ":src_trace_processor_plugins_structural_tree_partition_tables",
@@ -4964,15 +4962,6 @@ perfetto_filegroup(
         "src/trace_processor/plugins/strace/strace_trace_parser.h",
         "src/trace_processor/plugins/strace/strace_trace_tokenizer.cc",
         "src/trace_processor/plugins/strace/strace_trace_tokenizer.h",
-    ],
-)
-
-# GN target: //src/trace_processor/plugins/strace:strace_line_parser
-perfetto_filegroup(
-    name = "src_trace_processor_plugins_strace_strace_line_parser",
-    srcs = [
-        "src/trace_processor/plugins/strace/strace_line_parser.cc",
-        "src/trace_processor/plugins/strace/strace_line_parser.h",
     ],
 )
 
@@ -11583,7 +11572,6 @@ perfetto_cc_library(
         ":src_trace_processor_plugins_stdlib_docs_tables",
         ":src_trace_processor_plugins_storage_tables_storage_tables",
         ":src_trace_processor_plugins_strace_strace",
-        ":src_trace_processor_plugins_strace_strace_line_parser",
         ":src_trace_processor_plugins_string_functions_string_functions",
         ":src_trace_processor_plugins_structural_tree_partition_structural_tree_partition",
         ":src_trace_processor_plugins_structural_tree_partition_tables",
@@ -11923,7 +11911,6 @@ perfetto_cc_binary(
         ":src_trace_processor_plugins_stdlib_docs_tables",
         ":src_trace_processor_plugins_storage_tables_storage_tables",
         ":src_trace_processor_plugins_strace_strace",
-        ":src_trace_processor_plugins_strace_strace_line_parser",
         ":src_trace_processor_plugins_string_functions_string_functions",
         ":src_trace_processor_plugins_structural_tree_partition_structural_tree_partition",
         ":src_trace_processor_plugins_structural_tree_partition_tables",
