+++ test/trace_processor/diff_tests/parser/strace/basic.strace
@@ -1,3 +1,8 @@
-14:32:01.000000 openat(AT_FDCWD, "/etc/passwd", O_RDONLY) = 3
-14:32:01.000100 read(3, "root:x:0:0"..., 1024) = 312
-14:32:01.000200 close(3)                = 0
+1700000052.321000 openat(AT_FDCWD, "/etc/passwd", O_RDONLY) = 3
+1700000052.321100 read(3, "root:x:0:0"..., 1024) = 312
+1700000052.321200 close(3)                = 0
+1700000052.321300 openat(AT_FDCWD, "/nope", O_RDONLY) = -1 ENOENT (No such file or directory)
+--- SIGCHLD {si_signo=SIGCHLD, si_code=CLD_EXITED} ---
+1234 1700000052.321400 write(1, "hi", 2) = 2
++++ exited with 0 +++
+this line is not valid strace output and should be skipped
