+++ test/trace_processor/diff_tests/parser/strace/unfinished_resumed.strace
@@ -1,3 +1,3 @@
-14:32:01.000000 futex(0x7f, FUTEX_WAIT, 2, NULL <unfinished ...>
-14:32:01.000500 <... futex resumed> )  = 0
-14:32:01.000600 write(1, "done", 4)    = 4
+1700000052.321000 futex(0x7f, FUTEX_WAIT, 2, NULL <unfinished ...>
+1700000052.321500 <... futex resumed> )  = 0
+1700000052.321600 write(1, "done", 4)    = 4
